重構程式
 - 重構登入頁面
 - 重構註冊頁面
 - 嘗試使用前後分離技術
 - 使用api認證技術

+++ app/Http/Controllers/Api/ProductApiController.php
@@ -6,8 +6,10 @@ use App\Exceptions\APIException;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Services\ProductService;
+use Illuminate\Auth\Access\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 use Tymon\JWTAuth\Facades\JWTAuth;
 
@@ -19,11 +21,6 @@ class ProductApiController extends Controller
     private $service;
 
     /**
-    * @var JWTAuth
-    */
-    protected $user;
-
-    /**
      * Create a new controller instance.
      *
      * @return void
@@ -31,12 +28,17 @@ class ProductApiController extends Controller
     public function __construct(ProductService $service)
     {
         $this->service = $service;
-        $this->user = JWTAuth::parseToken()->authenticate();
+        $this->middleware('auth:api', ['except' => ['index', 'show']]);
     }
 
     public function index()
     {
         return $this->service->getProducts();
+    }
+
+    public function show($productId)
+    {
+        return $this->service->getProductById($productId);
     }
 
     public function store(Request $request)
@@ -52,7 +54,12 @@ class ProductApiController extends Controller
         ]);
 
         if ($validator->fails()) {
-            throw new APIException('驗證錯誤', 422);
+            //$messages = $validator->errors()->getMessages();
+            abort(422, '驗證錯誤');
+        }
+
+        if (Gate::allows('user')) {
+            abort(403, '權限錯誤');
         }
 
         $productForm = [
@@ -64,7 +71,6 @@ class ProductApiController extends Controller
             'description' => $request->get('description') ?? null,
             'content' => trim($request->get('content')) ?? '',
             'quantity' => intval($request->get('quantity')),
-            'imageUrl' => $request->get('imageUrl') ?? null,
             'size' => $request->get('size'),
         ];
 
@@ -72,38 +78,47 @@ class ProductApiController extends Controller
             $fileName = pathinfo($request->file('image')->getClientOriginalName(), PATHINFO_FILENAME);
             $extension = $request->image->extension();
             $imageName = time() . '.' . $extension;
-            $path = 'images';
-            $request->image->storeAs($path, $imageName, 'public');
+            $path = 'natz/images/';
+            Storage::disk('s3')->put($path . $imageName, file_get_contents($request->image));
+            $s3_path = Storage::disk('s3')->url('natz/images/' . $imageName);
 
-            $productForm['image'] = $path . '/' . $imageName;
+            $productForm['imageUrl'] = $s3_path;
+            $productForm['image'] = parse_url($s3_path, PHP_URL_PATH);
+        } elseif ($request->has('imageUrl')) {
+            $productForm['imageUrl'] = $request->get('imageUrl');
+            $productForm['image'] = null;
         }
 
         $product = Product::create($productForm);
 
         return response([
-            'product' => $product,
             'success' => true,
+            'product' => $product,
         ], 200);
     }
 
     public function update(Request $request, $productId)
     {
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|string|max:100',
-            'price' => 'required|string|max:10',
-            'quantity' => 'required|string|min:1',
-            'image' => 'mimes:jpeg,jpg,png,gif,svg|max:10000',
-            'category' => 'required|string|max:30',
-            'origin_price' => 'required|string|max:10',
-            'unit' => 'required|string|max:15',
-        ]);
+        try {
+            $request->validate([
+                'name' => 'required|string|max:100',
+                'price' => 'required|string|max:10',
+                'quantity' => 'required|string|min:1',
+                'image' => 'mimes:jpeg,jpg,png,gif,svg|max:10000',
+                'category' => 'required|string|max:30',
+                'origin_price' => 'required|string|max:10',
+                'unit' => 'required|string|max:15',
+            ]);
+        } catch (\Exception $e) {
+            abort(422, '驗證錯誤');
+        }
 
-        if ($validator->fails()) {
-            throw new APIException('驗證錯誤', 422);
+        if (Gate::allows('user')) {
+            abort(403, '權限錯誤');
         }
 
         if (! $product = Product::find($productId)) {
-            throw new APIException('查無產品', 404);
+            abort(404);
         }
 
         $productForm = [
@@ -115,38 +130,77 @@ class ProductApiController extends Controller
             'description' => $request->get('description') ?? null,
             'content' => trim($request->get('content')) ?? '',
             'quantity' => intval($request->get('quantity')),
-            'imageUrl' => $request->get('imageUrl') ?? null,
             'size' => $request->get('size'),
         ];
 
         if ($request->has('image')) {
-            if ($product['image'] !== 'img/noimage.jpeg') {
-                File::delete('storage/' . $product['image']);
+            if ($product['image'] !== '/natz/images/noimage.jpeg') {
+                Storage::disk('s3')->delete($product['image']);
             }
             $fileName = pathinfo($request->file('image')->getClientOriginalName(), PATHINFO_FILENAME);
             $extension = $request->image->extension();
             $imageName = time() . '.' . $extension;
-            $path = 'images';
-            $request->image->storeAs($path, $imageName, 'public');
+            $path = 'natz/images/';
+            Storage::disk('s3')->put($path . $imageName, file_get_contents($request->image));
+            $s3_path = Storage::disk('s3')->url('natz/images/' . $imageName);
 
-            $productForm['image'] = $path . '/' . $imageName;
+            $productForm['imageUrl'] = $s3_path;
+            $productForm['image'] = parse_url($s3_path, PHP_URL_PATH);
+        } elseif ($request->has('imageUrl')) {
+            if (
+                $product['image'] !== '/natz/images/noimage.jpeg'
+                && $request->get('imageUrl') !== $product['imageUrl']
+            ) {
+                Storage::disk('s3')->delete($product['image']);
+            }
+
+            $productForm['imageUrl'] = $request->get('imageUrl');
+            $productForm['image'] = null;
         }
 
         $status = $product->update($productForm);
-        return ['message' => $status];
+
+        if ($item = \Cart::session(auth()->id())->get($productId)) {
+            \Cart::session(auth()->id())->update($productId, array(
+                'name' => $product->name,
+                'price' => $product->price,
+                'quantity' => 1,
+                'attributes' => array(
+                    'imageUrl' => $product->imageUrl,
+                    'image' => $product->image,
+                    'unit' => $product->unit,
+                    'size' => $request->get('size'),
+                ),
+            ));
+        }
+
+        return response([
+            'success' => $status,
+            'message' => '產品已更新',
+            'product' => $product,
+        ], 200);
     }
 
     public function destroy($productId)
     {
-        try {
-            $product = Product::find($productId);
-        } catch (\Exception $e) {
-            throw new APIException('查無對應產品', 404);
+        if (Gate::allows('user')) {
+            abort(403, '權限錯誤');
         }
 
-        \Cart::session($this->user->id)->remove($productId);
+        if (! $product = Product::find($productId)) {
+            abort(404);
+        }
+
+        if ($product['image'] !== '/natz/images/noimage.jpeg' && $product['image'] != null) {
+            Storage::disk('s3')->delete($product['image']);
+        }
+
+        \Cart::session(auth()->id())->remove($productId);
 
         $status = $product->delete();
-        return ['message' => $status];
+        return response([
+            'success' => $status,
+            'message' => '產品已刪除',
+        ], 200);
     }
 }
